picasso  refactor for reuse
